refactor(mobile): extract the shared draft pipeline used by Send and WC
prepareSendDraft and composeSafeTxDraft duplicated the whole preview-and-stash
tail (hash, CGW /preview, synthesizeDraftTxDetails, setDraft) plus an identical
getVerifiedSafeSDK helper. Consolidate both into src/services/tx/draft.ts; the
two callers keep only their flow-specific composition (token-transfer params vs
dApp calls / CreateCall mapping). Covered by the existing suites of both
callers, which exercise the shared module directly.
